InfoPanel: split fs/mem helpers by OS so BSD/darwin/illumos build
CI failed on dragonfly / freebsd / darwin / illumos because the
initial fs_info_unix.go / mem_info_unix.go used a `!windows` build
tag but referenced Linux-only APIs — `syscall.Sysinfo(_t)` and the
`Statfs_t.Namelen` field name.

Reorganise to make the module portable:

* fs_info.go / mem_info.go — struct definitions, no build tag, so
  every platform sees the same shape.
* fs_info_linux.go / mem_info_linux.go — Linux implementations,
  now tagged `linux` (was `!windows`).
* fs_info_other.go / mem_info_other.go — `!linux && !windows` stubs
  returning ok=false. The info panel already skips sections when
  the platform can't fill them, so BSD/darwin/illumos users see a
  panel with just Computer / User / Current directory + whatever
  the panel can compute without OS-specific plumbing. Follow-up can
  add per-BSD ports if anyone wants them.

Local cross-builds pass for linux, darwin, freebsd, dragonfly,
netbsd, openbsd, illumos and windows on amd64.
